FilterButtons are displayed besides each other + split admin into more managable components
